Add Class window
Creates a csv file with course info. 1 csv file per course. These csv files will later be used to store information about the assignments for the course.
We should have a different folder for each semester, then each folder will have all of the course.csv files in them. If we do this, I'd need to change the directory that this file points at.