read extended description
The add current/previous semester buttons work now. They close the main page and take you to the next window. I couldn't figure out how to leave the window up without bugs. It should looks like the class buttons are gone, because I made it so now there will be another class button for each class you have in the current semester. So every time you add a class to the current semester, a new class button will appear on the main page.

As for the adjust semester button, I think we were planning on making that take you to a window that presents you with 3 buttons: add class, edit class, remove class. Those 3 buttons would then take you to other windows to add classes to the semester.

One important backend thing: there's a big block of code from line 47 to 112 (give or take) which finds the name of the file for the current semester (or most recent semester if the actual current semester has not yet been inputted).