read extended description
Added back button and adjusted window size

IMPORTANT: I think the add/edit assignments shouldn't be in the adjust current semester window. The add/edit assignments window should be a button in the cumulative rundown window when you click on one of those left-hand course buttons in the main page.