fixed main page --> class manager bug
if you added multiple classes and you tried to click a course button to open the main page, it used to go to the last course you added. This meant that your other courses were inaccessible. Now it brings you to the right course.

I also made the add/edit/remove assignment bring you back to the class manager page whenever you completed that function because it was being weird before.